Major bugfixes.
1) Fixed a bug that prevented directories on the server to be properly
deleted.

2) Fixed a bug that would cause files and directories to be deleted
often, and then re-uploaded. This was due to a missing comparison.

3) Serialized the directory creation, making sure that it was successful
before trying to create the next one, which may have been a
subdirectory.

4) Fixed the issue with the GUI freezing when it was waiting for a sync
to happen if the user tried to quit.

5) Fixed an issue that if a directory existed with the same Remote or
Local directory name, weird stuff would happen! This was due to an
incorrect comparison.

6) Fixed the issue where the system tray icon would not change when a
conflict occurred. (Apparently I had accidentally removed this when I
made the change to multiple accounts).

7) Added two new buttons to clear the log and clear & save the log.
Useful when debugging :) But useful in general, I suppose.

Possibly some other fixes, but now I forget them exactly. Probably just
some lost pointers and the like that I found :)